Split the cpp of the classes to be more clear
